infrastructure/function_code_previous/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions            # Azure Functions runtime support for Python
azure-identity             # Azure authentication via Managed Identity, CLI, or environment
azure-mgmt-containerservice # Manage Azure Kubernetes Service (AKS) clusters
azure-mgmt-compute         # Manage Azure Compute resources, including snapshots
kubernetes                 # Kubernetes Python client for interacting with cluster resources
PyYAML                     # Parse and handle YAML files (e.g., kubeconfig)

